remove duped cypress tests. env vars to indicate if screenshots should be taken or not
